+++ CHANGES.md
@@ -9,6 +9,7 @@
 - **ENH**: `get_mean_sun_angles` and `default_transform` are now cached (using `@cache`)
 - **ENH: Removing SNAP from Sentinel-3 pre-process -> Freeing optical data from SNAP dependency !**
 - **ENH: Enabling the use of other S3-SLSTR suffixes than `an` (stripe A at nadir position)**
+- **ENH: Archived Sentinel-3 products can now be processed**
 - **ENH: Removing raw `gdaldem` CLI from EOReader (the `HILLSHADE` and `SLOPE` bands are now slightly different !)** #10
 - **ENH: `HILLSHADE` is given in `float32` instead of `uint8`**
 - **ENH: Better management of `fspath` for cloud-stored products (download the files only once)**